Add separate header/implementation files for the classes in input_data.h/cpp.
