+++ python/boomer/common/cpp/input/nominal_feature_mask_dok.cpp
@@ -1,0 +1,10 @@
+#include "nominal_feature_mask_dok.h"
+
+
+bool DokNominalFeatureMask::isNominal(uint32 featureIndex) const {
+    return vector_.getValue(featureIndex);
+}
+
+void DokNominalFeatureMask::setNominal(uint32 featureIndex) {
+    vector_.setValue(featureIndex);
+}
