+++ python/boomer/seco/cpp/statistics_label_wise.h
@@ -7,7 +7,7 @@
  */
 #pragma once
 
-#include "../../common/cpp/input_data.h"
+#include "../../common/cpp/input/label_matrix.h"
 #include "../../common/cpp/statistics.h"
 #include "rule_evaluation_label_wise.h"
 #include "statistics.h"
