+++ python/setup.py
@@ -21,8 +21,13 @@ sources = [
     'boomer/common/cpp/data/vector_sparse_array.cpp',
     'boomer/common/cpp/indices/index_vector_full.cpp',
     'boomer/common/cpp/indices/index_vector_partial.cpp',
+    'boomer/common/cpp/input/feature_matrix_csc.cpp',
+    'boomer/common/cpp/input/feature_matrix_dense.cpp',
+    'boomer/common/cpp/input/feature_vector.cpp',
+    'boomer/common/cpp/input/label_matrix_dense.cpp',
+    'boomer/common/cpp/input/label_matrix_dok.cpp',
+    'boomer/common/cpp/input/nominal_feature_mask_dok.cpp',
     'boomer/common/cpp/random.cpp',
-    'boomer/common/cpp/input_data.cpp',
     'boomer/common/cpp/predictions.cpp',
     'boomer/common/cpp/rules.cpp',
     'boomer/common/cpp/post_processing.cpp',
